to test locally for eulerian flow around cylinder

+++ tests/2d_examples/test_2d_ball_shell_collision/ball_shell_collision.cpp
@@ -138,7 +138,7 @@ int main(int ac, char *av[])
 		//----------------------------------------------------------------------
 		//	From here iteration for particle relaxation begins.
 		//----------------------------------------------------------------------
-		int relax_step = 1000;
+		size_t relax_step = 1000;
 		while (sph_system.TotalSteps() < relax_step)
 		{
 			ball_relaxation_step_inner.parallel_exec();
